fix: Do not revert block proposal if vote fails and viceversa (#15613)
We were calling multicall3 with `allowFailure: false` for each request,
which means that the entire tx reverted if any of the aggregated calls
failed. This meant that, if block proposal failed, the vote would be
reverted, and viceversa. This PR sets `allowFailure` to true.

This change means we can no longer rely on the receipt `status` to know
if an action succeeded or not. So we overload the old `onResult`
callback for L1 requests, and have it inspect the tx logs to determine
if the action succeeded or not.