chore: begin txe contract call refactor (#15784)
In figuring out the strategy to pull off the txe oracle refactor, I
concluded that what makes the most sense first is to fully migrate to
the new way to call public, private and utility functions. Once done, we
should be able to eliminate large parts of txe_oracle (e.g.
`avmOpcodeCall`, `enqueuePublicCall`, `callPrivateFunction`), which will
make the following steps much simpler.

This is unfortunately easier said than done. This PR contains some of
the 'simple' changes - there's some nastier stuff that is not trivial to
get right that'll come later. I am being liberal in introducing changes
to the new `TestEnvironment` methods to try not to be too tied to one
specific API and iterate over it as I go.

Couple interesting points:
- we have multiple ways to call (most notably the `_void` variants)
mostly because the `CallInterface`s are unnecessarily complicated. We'll
need to refactor those to address this unfortunately, so that'll have to
wait.
- doing public calls through the processor lost us the capacity to
recover the revert reason. The public execution code is quite
convoluted, and I don't really have the capacity to go through it to
address this. I'm sure we can restore the behavior in the near future,
and given the huge mental context I'd first push forward these changes
and then worry about failure reasons.

The next things to look at are private calls and utility, notably how
those interact with notes that were just created, and authwits. Neither
seems to quite work, but I've not yet found out why. There's compounded
complexity in that many of the tests are not fantastic to begin with and
abuse very low level mechanics to which they should not have access, so
I find myself want to rewrite tests rather than port them. Hopefully
with this initial batch out of the way I can be more surgical about
those.

---------

Co-authored-by: benesjan <janbenes1234@gmail.com>