chore!: Nuke flushTxs (#15932)
Removes the flushTxs method in favor of setting minTxsPerBlock to zero
until the given condition is reached. We are not using this method in
production at all, just for testing. And it was flaky due to race
conditions: see [this run](http://ci.aztec-labs.com/7354b98827d088ca)
where a cross chain tx caused the entire test to time out since no
blocks were built, as the flushing failed, and the test got stuck in
`Not enough txs to build block`.

This change required adding a `getConfig` method so we could fetch the
original value for minTxsPerBlock so we could restore it afterwards.