chore: Clean up public input propagation in `acir_format` (#16411)
This PR restructures how we propagate public inputs in `acir_format`.
There is no change to the circuits/vks, the code is only restructured to
make it simpler too read.

This PR is in preparation for the introduction of recursive ClientIVC
verification in noir, as the new structure of the code will make it
easier to propagate the public inputs that come from recursive ClientIVC
verification.