feat!: aztec-nargo now directly exposes nargo. aztec-postprocess-contract transpiles and generates vks (#16172)
## What's done and not done?
1. ✅ Make aztec-nargo just be a direct exposure of the version of nargo
in the container. To this end aztec-nargo should from the users
experience, just be plain vanilla nargo, but just versioned as we need.
1. ✅ Fix the version number in aztecs nargo build to just be whatever
our current noir-ref is, maybe with a -aztec suffix or something to
ensure we have more context.
- Setting the `GIT_COMMIT` env var to the current noir-ref with `-aztec`
appended.
        ```
        % aztec-nargo --version
        nargo version = 1.0.0-beta.9
        noirc version = 1.0.0-beta.9+nightly-2025-08-04-aztec
        (git version hash: nightly-2025-08-04-aztec, is dirty: false)
        ```
- **Note that we are appending `-aztec` to `GIT_COMMIT`, which does not
end up in the "nargo version", but appears in the noirc version and git
version hash.** I don't see a clean/simple way to append `-aztec` to the
reported nargo version.
1. ✅ Put the current compile_and_postprocess script behind a command
e.g. `aztec-postprocess-contract` , make it no longer run nargo, but
rather just transform the artifact with the transpiler and vk
generation.
1. **[TODO]** ⚠️ Ensure devrel update docs promptly to reflect the added
compile step for aztec.

## An important note on caching and recompiling
1. The transpiler won't re-transpile a contract that is already
transpiled. *Yay!*
3. VK caching is still used. *Yay!*
4. Buuuuut... we can no longer use automatically usee nargo's
`--show-artifact-paths` to inform what contracts have actually changed
and therefore must be post-processed. So, the "update the artifact
sequentially with each generated vk" phase runs for every artifact no
matter what. If the source hasn't changed, the result will be the same,
but the files will be touched.