fix: address 142 (#16501)
Currently the rollup don't need to have the registry in its config, and
if it did and were to be updatable seems like it poses the same issue.

I think the sanest solution here will be to take the alternative route
and ensure that governance can simple claw back the funds. However, I
don't think that gov need to be constrained on where they can put them
for the simple reason that they already have the power of the issuer, so
they could already make funds if they wanted to.