feat: Enable the use of multiple EOAs (#16272)
This PR integrates the node key store and enables the use of multiple
EOAs for publishing to L1.

The primary changes are:

1. `SequencerPublisher` is now stateless. Slashing state has been moved
into the `Sequencer`. A new publisher is created every block using an
appropriate publisher account.
2. Publisher accounts and values of `coinbase` and `feeRecipient` are
selected with every new block and come from the key store.
3. L1TxUtils now tracks some state to determine where the account
can/should be used.
4. Signing of all transactions is no longer performed by viem's wallet
client, instead it is deferred to the key store allowing abstractions
such as remote signers.