chore(avm): remove trace truncation (#16609)
Given that Execution should never throw, every exception coming from
there will result in an unprovable transaction (crash).

I created a `TxExecutionException` for handleable TX-level errors. We
could've avoided exceptions altogether but the semantics of a try-catch
lets us skip the rest of the code when throwing and that makes it easier
to read, I think.