chore: greatly restructure txe's internals (#16903)
This got a bit out of hand in terms of size unfortunately - at this
stage it was hard to do part of the thing without being forced to do
pretty much the entire thing.

The largest change by far is the removal of large parts of `TXE`
(`txe_oracle.ts`). It has essentially been split into
`TXEOracleTopLevelContext`, which deals with all of the oracles
available 'at the top' of a test (i.e. outside of e.g. an
`env.private_context` call), and `TXE`, which remains the handler for
`private_context` and `utility_context` (merged for now since the split
is not easy - this is related to the issue described in
https://github.com/AztecProtocol/aztec-packages/issues/16800 re. the
implementations diverging). This means that the TXE session model is now
fully working, with each state being handled by the corresponding
handler, with state being updated on state transitions, etc.

Other than some minor cleanup, what remains is a) improving the awful
`privateCallNewFlow` (and public) in the top level context, and b)
splitting the utility and private contexts. But at least most of the
weird abuse of state, mixed interpretation of values like `blockNumber`
etc has gone away.

On top of that, many oracles have been cleaned up to not receive
unnecessary params, abuse certain features like the execution cache,
some obsolete ones have been deleted, etc.