chore(avm)!: gas gt migrations optimized (#16989)
This PR re-introduces the reverted PR #16856 which caused a performance
regression in "Log derivative inverse commitments".

After the introduction of separate destination lookup selectors, this
original PR does not affect the performance.

We introduce a specific gt selectors for the gas accounting (gt.sel_gas)
while for the specific gt lookups involved by the CALL opcode, we used
the already existing gt.sel_others selector. The reason is that we
expect much less gt events produced by the latter one.

Benchmarks on mainframe with 16 cores did not lead to any significant
performance changes (ca. 2.5 seconds for the derivative inverse
commitments).