feat: Soft-delete mined txs from tx pool (#17131)
When deleting mined txs due to a reorg or eviction, we soft-delete them
instead, and permanently delete them only when the original block number
in which they were mined gets finalized.

This allows us to keep txs from pruned epochs available in the pool,
which we need for identifying data withholding slash conditions in the
epoch prune watcher. Otherwise, since we were potentially deleting mined
txs from pruned epochs (eg if they became invalid as they were built on
top of removed blocks), we would lose them and incorrectly think we were
seeing a data withholding slash condition.

Other changes on this PR:
- Fixed the `sequencer-publisher` last-signal tracking. Since the
publisher is no longer long-lived, the `myLastSignals` cache was being
lost. It's now expanded to track every action, and kept alive by the
publisher factory.
- Added config var for the publisher factory to allow reusing publishers
in an invalid state, to reduce the flakes produced by timed out L1 txs,
which happen a lot when we time warp.
- Removed unused `setTxHash` to the `tx` object
- Minor logging fixes