chore: Try deflake l1 to l2 test (#17251)
We had a race condition in advanceBlock. There we checked if the block
before/after sending a tx and waiting for it to be mined was increased
by 1. But if we checked the block before while a block was being built
for another reason, we would end up with a bigger difference.

See failed run [here](http://ci.aztec-labs.com/c8b819fe94f98d46)