refactor: replace txe oracle in favor of pxe oracle (#17315)
This fully deletes the `TXE` class (from `txe_oracle`). This means that
`env.utility_context` and `env.private_context` are now running _the
exact same code_ as real simulations using PXE. TXE's job is now much
smaller: it just coordinates other components, but it doesn't have a lot
of behavior of its own.

Because in Noir tests we don't allow for nested private calls, I made
the `simulator` parameter in the oracle optional and simply throw an
error if we attempt to call it (which we don't in txe because the
translator rejects the call).

I went through the TXE oracles one by one, spotting how they matched or
differred from the PXE ones, and the only relevant difference I found is
that in TXE we used to immediately throw if the notify nullifier was
called with a settled nullifier - which PXE does not do. Both approaches
would make sense, and we want to have tests to make sure this behavior
is stable, but the important thing to note is that from now on _both TXE
and PXE_ will behave the same in this regard. We should try to maintain
this consistency as much as possible to avoid TXE from growing into an
unmaintainable useless monster. At any rate, this specific thing is
handled in #17317.

Once this PR goes in, and the TXE tests indirectly become (to some
extent) PXE tests as well, I'll begin opening PRs fixing the different
issues I've been finding.