feat!: entrypoint cleanup (#17422)
This PR renames all `...MethodOptions` to `...InteractionOptions`. These
are the collection of options that allow turning an `Interaction` into
an `ExecutionPayload` the wallet can actually process. Also the
convenience methods used to convert between the two are renamed from
`sanitize...Option` to `to...Option`.

This PR introduces the concept of specialized entrypoints for different
contracts. Our account contracts provide options to pay for the fee
themselves via fee juice, or rely on an external method. This was
previously achieved by a combination of lower level flags (`endSetup`,
`isFeePayer`). This new system avoids inconsistent combinations.

Finally, the `FeeJuicePaymentMethod` has been removed, since it wasn't
usable by apps and with the new flags the wallet can just command an
account contract (that supports it) to pay for the fee with its own
balance.