fix: Handle not-mined states in L1 publishers (#17412)
Packs several changes to the `L1TxUtils`:

- All timeouts are now computed wrt L1 timestamp, not local time. This
allows us to better control how many L1 slots without being mined we
want to wait before eg speeding up a tx. Before, the speed up would
trigger halfway through an L1 slot rather than at the beginning, wasting
valuable seconds.
- Adds a `txCancellationFinalTimeoutMs` setting. If a cancellation tx is
still not mined after all its speed up attempts, we give up on the tx
after this many ms, and switch to `NOT_MINED` state. This also resets
the nonce manager, so the publisher will recompute the nonce, re-using
the same nonce if the tx was dropped.
- Adds a `txUnseenConsideredDroppedMs`. If a cancellation tx is not seen
in the mempool for this many ms, we'll consider it was just dropped and
happily revert back to `IDLE`.
- Fixes the order in which we check the tx state: we check if the tx is
mined _before_ checking if it is expired and attempting to cancel it.
- Adds retries to fetching receipts.
- Handles timeouts where `cancelTxOnTimeout` is false, which would leave
the publisher in a `SPEED_UP` state.
- Does not fire a cancellation tx if the publisher was interrupted (ie
the node is being stopped), or if the tx was dropped from the mempool
(no need to cancel it if it was forgotten).
- Removes no longer used env vars `PUBLISH_RETRY_INTERVAL` and
`TX_PROPAGATION_MAX_QUERY_ATTEMPTS`.

Given the changes from #17378 and this PR, updates the publisher manager
so it defaults to always return a publisher, even on busy states, and
adds an e2e publisher test to check that we can handle this.

Aside from the changes to L1 publishers, also adds a few more eth cheat
codes, in particular `mineEmptyBlock` which removes all txs from the
mempool, mines a block, and then re-adds them immediately after.