feat: add ts to docs example code (#17627)
I did some re-arranging of the structure as well, to make this play a
bit nicer.

The main flow is that on CI / PR, bootstrap.sh in docs is run, which
should run bootstrap.sh in examples (which compiles the noir contracts,
and runs bootstrap.sh in examples/ts. examples/ts defines a common setup
for typescript projects, including a common typescript config, as well
as defines a way that dependencies are installed from each individual
project.

It sets up a yarn project from the project folder, installs
dependencies, and runs tsc to type check the files to make sure breaking
interface changes have not caused any compilation issues.

This can be extended in the future to run actual tests.

I used to have a setup.sh for each project, but decided against it to
remove unnecessary duplication of dependency installation. The benefit
of this is deduping logic / code, at the downside of having the exact
same bash setup between the tutorial and this workflow.