feat!: Ecdsa audit part 8 (#17813)
### 🧾 Audit Context

Refactored ECDSA constraints to handle `predicate`, remove duplicated
constraints, and refactor testing.

### 🛠️ Changes Made

If `predicate` is witness false, then we:
- Override the first byte of `r` and `s` (components of the signature)
and make them equal to `1`. This ensures that the checks `0 < r <
Fr::modulus` and `0 < s < Fr::modulus/2` pass. Note that his only works
when `Fr::modulus/2 > 2^241`, which is the case for secp256k1 and
secp256r1.
- Override the public key to `2 * G`, where `G` is the generator of the
curve.
- Check the ecdsa verification result against itself to be sure that
there are no failing constraints

This PR closes
https://github.com/AztecProtocol/barretenberg/issues/1562: we remove the
duplicate check that the x and y coordinate of a public key are smaller
than the base field modulus

We refactor tests to use the new interface `TestClassWithPredicate`.

There is a change in the VKs because of the changes in
`ecdsa_constraint.cpp`.

### ✅ Checklist

- [ ] Audited all methods of the relevant module/class
- [ ] Audited the interface of the module/class with other (relevant)
components
- [X] Documented existing functionality and any changes made (as per
Doxygen requirements)
- [X] Resolved and/or closed all issues/TODOs pertaining to the audited
files
- [ ] Confirmed and documented any security or other issues found (if
applicable)
- [X] Verified that tests cover all critical paths (and added tests if
necessary)
- [ ] Updated audit tracking for the files audited (check the start of
each file you audited)

### 📌 Notes for Reviewers