refactor!: renaming `#[internal]` as `#[only_self]` + updated tests (#17958)
Closes
https://linear.app/aztec-labs/issue/F-20/rename-internal-to-something-else-eg-externalprivate-only-self

We want for internal to mean the same thing as in Solidity where
internal function can be called only from the same contract and is also
inlined (EVM JUMP opcode and not EVM CALL). The original implementation
of our `#[internal]` macro also results in the function being callable
only from the same contract but it results in a different call (hence it
doesn't map to EVM JUMP). This is very confusing hence I am doing the
rename in this PR.

When checking whether the `#[only_self]` is tested I stumbled upon old
e2e tests doing that. Instead of updating them decided to move them over
to Noir.

A true `#[internal]` will be introduced in a followup PR.