feat!: execution payload refactor (#18338)
Part of:
https://linear.app/aztec-labs/issue/F-144/improve-interaction-composability

This PR moves `ExecutionPayload` to `stdlib`, cementing its role as the
minimal unit of execution on the network.

As a part of this, the `feePayer` for an execution payload (IF it
includes a payment method, that is) is now embedded in the execution
payload, since previously it was specified after the fact which was
confusing and prone to errors. Finally, renamed a few fee-handling
functions in the wallet as discussed with @benesjan in the past.