fix: make round size limit inclusive max (#18500)
In the tally slashing proposer there was a constraint on the
`ROUND_SIZE` to be smaller than the max value, while it should be
allowed be the less than or equal.

The limit is to making a config that could run out of bounds when it
tries to write the to the `RoundVotes`.