chore: Avoid capturing a mutable reference in `remove_constraints_if` (#18595)
This should unblock https://github.com/noir-lang/noir/pull/10497 to be
merged into Noir. Thanks to @asterite for pointing out this simple fix.

Passing mutable references from ACIR to Brillig is invalid, however it
slipped in by capturing `self` in a lambda, which is a `NoteEmission`
that contains a `context: &mut PrivateContext` field.